linux: Move `sctp.h` API to `src/new`
Link: https://github.com/torvalds/linux/blob/ab9de95c9cf952332ab79453b4b5d1bfca8e514f/include/uapi/linux/sctp.h
